Added documentation and fixed mutagen metadata bug
All inputted metadata had to be string, but the index was
an integer, so it had to be converted to a string to actually
be able to be placed in the metadata. Also, KeyboardInterrupt
allows the user more options instead of just skipping the video.
Users can now retry the same video, skip to next one, or exit.
Theres probably some small formatting changes, I dont remember.

+++ yt.py
@@ -1,17 +1,67 @@
+"""
+youtube-dl mp3 music download script
+
+This is a script that downloads youtube music videos as mp3 files.
+It is designed to be used in the same directory as the script
+because it will create a folder in that directory. Also, I'm too
+lazy to test out the edge cases of ensuring the functionality
+works in a different directory. Also, yes it downloads as mp3...
+but that is ONLY because youtube-dl only allows thumbnail embedding
+for mp3 and .ogg does not really support embedded images. Maybe it can
+idk I am too lazy.
+
+The purpose of creating this script as opposed to editing youtube-dl
+configs is to have a much better logging system (in my opinion -- at
+least its more customizable, okay?), to automatically retry failed
+downloads (for some reason, youtube-dl will just fail to download the
+.webm file. I think its because Google blocked quick sucessive downloads),
+and to have a better way to specify important metadata to me (mostly
+track numbers; youtube-dl still has not implemented this >:/).
+
+Args:
+    url: The URL of the video(s) to download
+    artist (optional): The artist of the video(s)
+    album (optional): The album of the video(s)
+    tries (optional): The number of times to retry downloading a video
+        if it fails. Defaults to 3.
+
+Returns:
+    Nothing really, it just downloads the video(s)
+
+Raises:
+    All kinds of exceptions idk read the code
+"""
+
+
 import os
 import getopt
 import sys
 
-import mutagen  # type: ignore
+from mutagen.easyid3 import EasyID3  # type: ignore
 import termcolor  # type: ignore
 import youtube_dl  # type: ignore
 
 
 class Logger(object):
-    """Use these methods to create manual log messages"""
+    """
+    Logger static class to make logging prettier and standardized
+    These methods are to create manual log messages by the programmer
+    """
+    @staticmethod
+    def prompt(message):
+        print(termcolor.colored("[PROMPT]", "blue"), message)
+
     @staticmethod
     def info(message):
         print(termcolor.colored("[INFO]", "green"), message)
+
+    @staticmethod
+    def success(message):
+        print(termcolor.colored("[SUCCESS]", "green"), message)
+
+    @staticmethod
+    def fail(message):
+        print(termcolor.colored("[FAIL]", "red"), message)
 
     @staticmethod
     def warn(message):
@@ -21,7 +71,7 @@ class Logger(object):
     def err(message):
         print(termcolor.colored("[ERROR]", "red"), message)
 
-    """These methods are to only be used and called by youtube_dl"""
+    """These methods here are to only be used and called by youtube_dl"""
     @staticmethod
     def hooks(h):
         if h["status"] == "finished":
@@ -70,6 +120,12 @@ youtubedl_options = {
 
 
 def parse_user_arguments():
+    """
+    Handle arguments when the user runs the script in the terminal
+    This just parases the arguments and stores them as variables to be
+    used later. Also, it determines if the download is a playlist or not.
+    """
+
     artist = None
     album = None
     url = None
@@ -87,7 +143,7 @@ def parse_user_arguments():
         Logger.err("Invalid arguments:", err)
         sys.exit(1)
 
-    # Parse arguments
+    # Parse arguments and store them in variables
     for argument, value in arguments:
         if argument in ("-h", "--help"):
             Logger.info("Downloads youtube music videos as mp3 files")
@@ -116,6 +172,13 @@ def parse_user_arguments():
 
 
 def remove_build_files(filename=None):
+    """
+    Removes the build files that youtube-dl creates. This can be the .webm
+    files that are downloaded, but it was intended for the .jpg thumbnail
+    files that youtube-dl does not clean up and does not seem to have the
+    option to do so, unlike the .webm files which can be cleaned up.
+    """
+
     # Assume that this is run in the same directory as the file
     possible_build_extensions = {
         ".jpg", ".jpeg", ".png", ".gif", ".bmp", ".webp",  # thumbnails
@@ -131,7 +194,7 @@ def remove_build_files(filename=None):
                 removed_build_files.append(file)
     else:
         for extension in possible_build_extensions:
-            if os.path.isfile(filename + extension):
+            if filename and os.path.isfile(filename + extension):
                 os.remove(filename + extension)
                 removed_build_files.append(filename + extension)
 
@@ -139,6 +202,15 @@ def remove_build_files(filename=None):
 
 
 def gather_video_info(artist, album, url, is_playlist):
+    """
+    Get every video's information for the metadata of the mp3 file and
+    (less importantly), the logs. This is also to separate each video
+    in a playlist (as opposed to downloading the entire playlist via
+    a single command by youtube-dl). It allows us to be more
+    meticulous and detailed with each video download (e.g. logs,
+    metadata, any other custom commands to run inbetween downloads, etc.).
+    """
+
     # Get info of the video(s)
     Logger.info("Starting to gather info about video(s)...")
     info = ytdl.extract_info(url, download=False)
@@ -192,9 +264,15 @@ def gather_video_info(artist, album, url, is_playlist):
 
 
 if __name__ == "__main__":
+    """
+    Main function that runs when the script is run in the terminal.
+    """
+
+    # Clear cache files because that apparently helps with less errors
     Logger.info("Removing youtube-dl cache files...")
     os.system("youtube-dl --quiet --rm-cache-dir")
 
+    # Apply user options when downloading the videos itself
     ytdl = youtube_dl.YoutubeDL(youtubedl_options)
 
     # url is a mandatory argument, we assume it is not None
@@ -220,8 +298,11 @@ if __name__ == "__main__":
             try:
                 # Download the video
                 Logger.info(f"Starting to download {video['title']}...")
-                Logger.info(f"Video {video['index']}/{len(videos)} | " +
-                            f"Try {tries}/{max_tries}")
+                if is_playlist:
+                    Logger.info(f"Video {video['index']}/{len(videos)} | " +
+                                f"Try {tries}/{max_tries}")
+                else:
+                    Logger.info(f"Try {tries}/{max_tries}")
                 ytdl.download([video['url']])
                 Logger.info(f"Finished downloading {video['title']}")
 
@@ -230,33 +311,68 @@ if __name__ == "__main__":
                 Logger.info(f"Removed {thumbnail_files} thumbnail")
 
                 # Add metadata to the mp3 files
-                audio_file = mutagen.easyid3.EasyID3(
-                    video["filename"] + ".mp3"
-                )
+                Logger.info("Adding metadata to mp3 files...")
+                audio_file = EasyID3(video["filename"] + ".mp3")
 
                 # Downloaded music may not be in album, hence the default Nones
-                # But all music should have a title and an artist
+                # But we assmue all music should have a title and an artist
                 if video["index"]:
-                    audio_file["tracknumber"] = video["index"]
+                    audio_file["tracknumber"] = str(video["index"])
                 if video["album"]:
                     audio_file["album"] = video["album"]
                     audio_file["albumartist"] = video["artist"]
                 audio_file["title"] = video["track"]
                 audio_file["artist"] = video["artist"]
                 audio_file.save()
+                Logger.info("Finished adding metadata to mp3 files")
+
+                Logger.success("Downloaded and processed " +
+                               f"{video['filename']}.mp3 successfully!")
 
                 break
             except KeyboardInterrupt:
-                Logger.warn("Keyboard interrupt. Skipping to next video...")
-                break
-            except:
-                Logger.err(f"Failed to download {video['title']}. " +
-                           "Trying again...")
+                # Keyboard interrupt pauses the program and allows
+                # the user to control the download: skip the video, cancel etc.
+                Logger.warn("Keyboard interrupt detected!")
+                Logger.warn("Pausing and waiting for user decision...")
+
+                while True:
+                    # Ensure the user inputs a valid option
+                    user_decision = input(
+                        Logger.prompt("Retry, Skip, or Exit? (r|s|x) "))
+                    if user_decision in {"r", "s", "x"}:
+                        break
+                    else:
+                        Logger.warn("Invalid input!")
+
+                # Handle user decision
+                if user_decision == "r":
+                    Logger.info("Trying again...")
+                    continue
+                elif user_decision == "s":
+                    Logger.info("Skipping this video...")
+                    break
+                elif user_decision == "x":
+                    Logger.info("Exiting...")
+                    sys.exit(0)
+
+            except Exception as e:
+                Logger.err(f"Failed to download/process {video['title']}. ")
+                Logger.err(e)
+                if tries <= max_tries:
+                    Logger.info("Trying again...")
+                else:
+                    # We have tried max_tries times and failed
+                    Logger.warn(f"Exceeded max tries of {max_tries}. " +
+                                "Skipping this video...")
+
+                # Clear the cache files to help with less errors
                 Logger.info("Removing youtube-dl cache files...")
                 os.system("youtube-dl --quiet --rm-cache-dir")
                 continue
 
+    # We are finished the downloads and its post processing!
     Logger.info("Finished downloading all videos")
-    Logger.info(f"Removed {remove_build_files()}")
+    Logger.info(f"Removed {remove_build_files()} remaining build files")
     Logger.info("Exiting...")
     sys.exit(0)
